Autres chiffres en double au lieu de int get sur Tower

+++ app/src/main/java/com/towerint/projectile.java
@@ -3,7 +3,7 @@ package com.towerint;
 abstract class projectile {
     private double x;
     private double y;
-    private int speed; // la vitesse à laquelle se déplace le projectile
-    private int power;
+    private double speed; // la vitesse à laquelle se déplace le projectile
+    private double power;
     //private image; TODO créer une classe pour l'image
 }
